[feat] - added skills to work page
