+++ src/app/work/page.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import { IBM_Plex_Sans } from 'next/font/google'
 import { Skills } from '@/components'
 
 // If loading a variable font, you don't need to specify the font weight
@@ -8,7 +7,9 @@ import { Skills } from '@/components'
 export default function Page() {
   return (
     <main className="bg-mint1">
-      <h1>Work</h1>
+      <h1 className="text-8xl text-center font-semibold text-mint12 py-8">
+        work
+      </h1>
       <Skills />
     </main>
   )
